+++ src/main/java/calculator/OutputView.java
@@ -1,7 +1,7 @@
 package calculator;
 
 public class OutputView {
-    void printResult(int result) {
+    void printResult(long result) {
         System.out.println("결과 : " + result);
     }
 }
